Moved GroupVisual from c21737525 to tudublin, since Luke and Hao will be using it as well
