+++ java/src/c21737525/VilimsSphere.java
@@ -1,5 +1,8 @@
 package c21737525;
 
+import ie.tudublin.*;
+import processing.core.*;
+
 public class VilimsSphere {
 
     GroupVisual gv;
